requirements.txt
flask==3.0.3
flask-socketio==5.3.6
python-nmap==0.7.1
requests==2.31.0
python-dotenv==1.0.1
fpdf2==2.7.9
flask-login==0.6.3
eventlet==0.36.1
flask-limiter==3.5.0
beautifulsoup4==4.14.3
dnspython==2.8.0
python-whois==0.9.6
